feat(inventory): add CRUD services and controllers for inventory management - Implement Ubicacion, InventarioItem and MovimientoInventario services with full CRUD operations - Add corresponding REST controllers with OpenAPI documentation - Enhance OrdenCompra and OrdenVenta services to include pending reception/dispatch queries - Update repository methods with pessimistic locking for concurrent inventory updates - Modify stock handling logic to automatically adjust inventory during order operations - Extend EnumCodigoEstado with inventory movement types and add utility methods

+++ stockflow-backend-web-api/src/main/java/com/stockflow/core/repository/MovimientoInventarioRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import com.stockflow.core.entity.MovimientoInventario;
 
 @Repository
-public interface MovimientoInventarioRepository extends JpaRepository<MovimientoInventario, Long> {
+public interface MovimientoInventarioRepository extends JpaRepository<MovimientoInventario, Integer> {
     
     @Query("SELECT FUNCTION('MONTH', m.fechaMovimiento) as mes, " +
            "SUM(CASE WHEN m.tipoMovimiento = 'ENTRADA' THEN m.cantidad ELSE 0 END) as totalEntradas, " +
